Remove nextrouter, port over some functionality to embed
